+++ gammapy/estimators/points/__init__.py
@@ -3,7 +3,7 @@ from .core import FluxPoints
 from .lightcurve import LightCurveEstimator
 from .profile import FluxProfileEstimator
 from .sed import FluxPointsEstimator
-from .sensitivity import SensitivityEstimator, ParameterSensitivityEstimator
+from .sensitivity import SensitivityEstimator
 
 __all__ = [
     "FluxPoints",
@@ -11,5 +11,4 @@ __all__ = [
     "FluxProfileEstimator",
     "LightCurveEstimator",
     "SensitivityEstimator",
-    "ParameterSensitivityEstimator",
 ]
